refactor: move agent execution enums to types to avoid circular imports
