+++ mypy.ini
@@ -12,7 +12,7 @@
 #
 
 [mypy]
-python_version = 3.10
+python_version = 3.11
 warn_unused_configs = True
 warn_redundant_casts = True
 show_error_codes = True
